+++ src/cli/utils/chains.js
@@ -1,5 +1,5 @@
 import Debug from 'debug';
-import { getChainDefaults } from '../../common/utils/config.js';
+import { getId, getChainDefaults } from '../../common/utils/config.js';
 import IExecContractsClient from '../../common/utils/IExecContractsClient.js';
 import { EnhancedWallet } from '../../common/utils/signers.js';
 import { loadChainConf } from './fs.js';
@@ -8,27 +8,21 @@ import { getReadOnlyProvider } from '../../common/utils/providers.js';
 
 const debug = Debug('iexec:chains');
 
-const CHAIN_ALIASES_MAP = {
-  1: 'mainnet',
-  134: 'bellecour',
-};
-
-const CHAIN_NAME_MAP = {
-  1: { id: '1' },
-  mainnet: { id: '1' },
-  134: { id: '134' },
-  bellecour: { id: '134' },
-};
-
 const createChainFromConf = (
   chainName,
   chainConf,
-  { bridgeConf, providerOptions, txOptions = {} } = {},
+  {
+    bridgeConf,
+    providerOptions,
+    txOptions = {},
+    allowExperimentalNetworks = false,
+  } = {},
 ) => {
   try {
     const chain = { ...chainConf };
     const provider = getReadOnlyProvider(chainConf.host, {
       providers: providerOptions,
+      allowExperimentalNetworks,
     });
 
     chain.name = chainName;
@@ -45,6 +39,7 @@ const createChainFromConf = (
       chain.bridgedNetwork = { ...bridgeConf };
       const bridgeProvider = getReadOnlyProvider(bridgeConf.host, {
         providers: providerOptions,
+        allowExperimentalNetworks,
       });
       chain.bridgedNetwork.contracts = new IExecContractsClient({
         provider: bridgeProvider,
@@ -63,26 +58,34 @@ const createChainFromConf = (
 };
 
 export const loadChain = async (
-  chainName,
+  chainNameOrId,
   { txOptions, spinner = Spinner() } = {},
 ) => {
   try {
     const chainsConf = await loadChainConf();
-    debug('chainsConf', chainsConf);
+    const { allowExperimentalNetworks } = chainsConf;
     const providerOptions = chainsConf.providers;
     let name;
     let loadedConf;
-    if (chainName) {
-      if (chainsConf.chains[chainName]) {
-        loadedConf = chainsConf.chains[chainName];
-        name = chainName;
+    if (chainNameOrId) {
+      if (chainsConf.chains[chainNameOrId]) {
+        loadedConf = chainsConf.chains[chainNameOrId];
+        name = chainNameOrId;
       } else {
-        const alias = CHAIN_ALIASES_MAP[chainName];
+        const { name: alias } = getChainDefaults(
+          getId(chainNameOrId, {
+            allowExperimentalNetworks,
+          }),
+          {
+            allowExperimentalNetworks,
+          },
+        );
         if (alias && chainsConf.chains[alias]) {
           loadedConf = chainsConf.chains[alias];
           name = alias;
         }
-        if (!name) throw Error(`Missing "${chainName}" chain in "chain.json"`);
+        if (!name)
+          throw Error(`Missing "${chainNameOrId}" chain in "chain.json"`);
       }
     } else if (chainsConf.default) {
       if (chainsConf.chains[chainsConf.default]) {
@@ -96,11 +99,16 @@ export const loadChain = async (
       throw Error('Missing chain parameter. Check your "chain.json" file');
 
     const idConf = {
-      ...CHAIN_NAME_MAP[name],
-      ...(loadedConf.id && { id: loadedConf.id }),
+      id:
+        loadedConf.id ||
+        getId(name, {
+          allowExperimentalNetworks,
+        }),
     };
 
-    const defaultConf = getChainDefaults(idConf);
+    const defaultConf = getChainDefaults(idConf.id, {
+      allowExperimentalNetworks,
+    });
 
     debug('loading chain', name);
     debug('loadedConf', loadedConf);
@@ -120,7 +128,14 @@ export const loadChain = async (
       if (chainsConf.chains[bridgedChainNameOrId]) {
         bridgeLoadedConf = chainsConf.chains[bridgedChainNameOrId];
       } else {
-        const alias = CHAIN_ALIASES_MAP[bridgedChainNameOrId];
+        const { name: alias } = getChainDefaults(
+          getId(bridgedChainNameOrId, {
+            allowExperimentalNetworks,
+          }),
+          {
+            allowExperimentalNetworks,
+          },
+        );
         if (alias && chainsConf.chains[alias]) {
           bridgeLoadedConf = chainsConf.chains[alias];
         }
@@ -128,10 +143,15 @@ export const loadChain = async (
           throw Error(`Missing "${name}" chain in "chain.json"`);
       }
       const bridgeIdConf = {
-        ...CHAIN_NAME_MAP[bridgedChainNameOrId],
-        ...(bridgeLoadedConf.id && { id: bridgeLoadedConf.id }),
+        id:
+          bridgeLoadedConf.id ||
+          getId(bridgedChainNameOrId, {
+            allowExperimentalNetworks,
+          }),
       };
-      const bridgeDefaultConf = getChainDefaults(bridgeIdConf);
+      const bridgeDefaultConf = getChainDefaults(bridgeIdConf.id, {
+        allowExperimentalNetworks,
+      });
       debug('bridgeLoadedConf', bridgeLoadedConf);
       debug('bridgeDefaultConf', defaultConf);
       bridgeConf = {
@@ -150,6 +170,7 @@ export const loadChain = async (
       bridgeConf,
       providerOptions,
       txOptions,
+      allowExperimentalNetworks,
     });
     spinner.info(`Using chain ${name} [chainId: ${chain.id}]`);
     return chain;