+++ src/lib/utils.d.ts
@@ -42,6 +42,12 @@ export const getSignerFromPrivateKey: (
      * providers options
      */
     providers: ProviderOptions;
+    /**
+     * if true allows using a provider connected to an experimental networks (default false)
+     *
+     * ⚠️ experimental networks are networks on which the iExec's stack is partially deployed, experimental networks can be subject to instabilities or discontinuity. Access is provided without warranties.
+     */
+    allowExperimentalNetworks?: boolean;
   },
 ) => EnhancedWallet;